pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyrecap"
version = "0.0.1"  # overridden by versioneer
description = "Scientific framework for Reproducible Experiment Capture and Provenance"
authors = [{ name = "Venkateswaran Shekar", email = "vshekar1@bnl.gov" }]
requires-python = ">=3.8"
readme = "README.rst"
license = { text = "BSD-3-Clause" }
dependencies = [
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "python-slugify"
]

[project.optional-dependencies]
dev = [
    "pytest"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["recap*"]

[tool.versioneer]
VCS = "git"
style = "pep440"
versionfile_source = "recap/_version.py"
versionfile_build = "recap/_version.py"
tag_prefix = ""
parentdir_prefix_version = "recap-"

[tool.hatch.build.targets.wheel]
packages = ["recap"]

[tool.black]
line-length = 115
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist

  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
)/
'''

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "C90"]
ignore = ["E203", "E501"]  # Black-style formatting norms

[tool.ruff.lint.per-file-ignores]
"**/versioneer.py" = ["ALL"]
"**/_version.py" = ["ALL"]
"**/setup.py" = ["ALL"]
"**/docs/**" = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = []

